change Dog age column from decimal to float

+++ db/migrate/20180424122609_create_dogs.rb
@@ -2,7 +2,7 @@ class CreateDogs < ActiveRecord::Migration[5.2]
   def change
     create_table :dogs do |t|
       t.string :name
-      t.decimal :age
+      t.float :age
       t.string :gender
       t.boolean :trained
       t.integer :fee
